+++ examples/core/core_3d_camera_first_person.cpp
@@ -61,7 +61,7 @@ int main() {
 
             background.ClearBackground();
 
-            camera.BeginMode3D();
+            camera.BeginMode();
 
                 DrawPlane(Vector3{ 0.0f, 0.0f, 0.0f }, Vector2{ 32.0f, 32.0f }, LIGHTGRAY); // Draw ground
                 DrawCube(Vector3{ -16.0f, 2.5f, 0.0f }, 1.0f, 5.0f, 32.0f, BLUE);     // Draw a blue wall