+++ include/ModelAnimation.hpp
@@ -80,7 +80,7 @@ class ModelAnimation : public ::ModelAnimation {
     /**
      * Update model animation pose
      */
-    inline ModelAnimation& Update(::Model model, int frame) {
+    inline ModelAnimation& Update(const ::Model& model, int frame) {
         ::UpdateModelAnimation(model, *this, frame);
         return *this;
     }
@@ -88,7 +88,7 @@ class ModelAnimation : public ::ModelAnimation {
     /**
      * Check model animation skeleton match
      */
-    inline bool IsValid(::Model model) const {
+    inline bool IsValid(const ::Model& model) const {
         return ::IsModelAnimationValid(model, *this);
     }