+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Mouse::SetX()` and `Mouse::SetY()`
 - Error correction when unloading images, materials, models and meshes
 
+### Changed
+- `Mouse` functions are now `static`. Use `Mouse::SetX()` instead of instantiating a `Mouse` object.
+- `Camera*::BeginMode()` and `Camera*::EndMode()` no longer have 2D/3D in the name
+
 ## [v3.5.0-alpha1] - 2020-12-26
 ### Changed
 - Update to raylib 3.5.0